# backend/Application/Organisers/Commands/RemoveContactInformation/RemoveContactInformationResponse.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Application.Organisers.Commands.RemoveContactInformation
{
    public class RemoveContactInformationResponse
    {
    }
}