# backend/Web/wwwroot/swagger/specification.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Loppe Portalen API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Organiser/new": {
      "post": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_CreateOrganiser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganiserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrganiserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organiser/add/contactInformation": {
      "post": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_AddContactInformation",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrganiserContactInformationRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddContactInformationResponse2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Organiser/{pageNumber}/{pageSize}": {
      "get": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_GetOrganisers",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrganisersWithPaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organiser/{id}": {
      "get": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_GetOrganiser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrganiserQueryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organiser/edit": {
      "put": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_EditOrganiser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditOrganiserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditOrganiserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organiser/contactInfo": {
      "delete": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_RemoveContactInformation",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveContactInformationRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveContactInformationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Organiser/banner/upload": {
      "put": {
        "tags": [
          "Organiser"
        ],
        "operationId": "Organiser_UploadOrganiserBanner",
        "parameters": [
          {
            "name": "organiserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBannerResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "MarketBaseVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "marketId": {
            "type": "integer",
            "format": "int32"
          },
          "marketName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "isCancelled": {
            "type": "boolean"
          },
          "totalStallCount": {
            "type": "integer",
            "format": "int32"
          },
          "availableStallCount": {
            "type": "integer",
            "format": "int32"
          },
          "occupiedStallCount": {
            "type": "integer",
            "format": "int32"
          },
          "categories": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OrganiserBaseVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "streetNumber": {
            "type": "string",
            "nullable": true
          },
          "appartment": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ContactInfoBaseVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ContactInfoType"
          }
        }
      },
      "ContactInfoType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "PHONE_NUMER",
          "FACEBOOK",
          "TWITTER",
          "EMAIL",
          "INSTAGRAM",
          "TIKTOK"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ]
      },
...
      "CreateOrganiserResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "appartment": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateOrganiserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "appartment": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AddContactInformationResponse2": {
        "type": "object",
        "additionalProperties": false
      },
      "AddOrganiserContactInformationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiserId": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ContactInfoType"
          }
        }
      },
...
      "GetOrganisersWithPaginationResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organisers": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PaginatedListOfOrganiser"
              }
            ]
          }
        }
      },
      "PaginatedListOfOrganiser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Organiser"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
      "Organiser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GetOrganiserQueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiser": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GetOrganiserVM"
              }
            ]
          }
        }
      },
      "GetOrganiserVM": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrganiserBaseVM"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "markets": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/MarketBaseVM"
                }
              },
              "contacts": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/ContactInfoBaseVM"
                }
              },
              "imageData": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "EditOrganiserResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "EditOrganiserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiserId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "appartment": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "RemoveContactInformationResponse": {
        "type": "object",
        "additionalProperties": false
      },
      "RemoveContactInformationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiserId": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UploadBannerResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiserId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}